added delete-method

+++ src/main/java/com/example/fobidb/controller/TeacherController.java
@@ -1,8 +1,10 @@
 package com.example.fobidb.controller;
 
 import com.example.fobidb.entity.Teacher;
+import com.example.fobidb.repository.TeacherRepository;
 import com.example.fobidb.service.TeacherService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,7 +14,7 @@ import java.util.Optional;
 /**
  ** @Author: Chris M.
  ** @Author: Mert Kaan K.
- ** @Author: Michael P.
+ ** @Author: Michel P.
  ** @Date: 07.04.2025
  *
  ** @Description: Diese Klasse verwaltet die HTTP-Anfragen für die Lehrer-Entität.
@@ -22,12 +24,14 @@ import java.util.Optional;
  ** @Reason: Kommentar hinzugefügt
  */
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(path = "api/v1/teacher")
 public class TeacherController {
     // Initialisiert den TeacherService
     private final TeacherService teacherService;
+    private final TeacherRepository teacherRepository;
 
     // Gibt eine Liste aller Lehrer zurück
     @GetMapping
@@ -53,9 +57,18 @@ public class TeacherController {
 
     // Löscht einen Lehrer
     @DeleteMapping
-    public ResponseEntity<Teacher> DeleteTeacher(@RequestBody Teacher teacher) {
-        teacherService.DeleteTeacher(teacher);
-        return ResponseEntity.ok(teacher);
+    public ResponseEntity<Teacher> DeleteTeacher(@RequestBody Long id) {
+        log.info("Deleting teacher with id: " + id);
+        Optional<Teacher> optionalTeacher = teacherRepository.findTeacherById(id);
+
+        if (optionalTeacher.isPresent()) {
+            Teacher teacher = optionalTeacher.get();
+            teacherService.DeleteTeacher(teacher);
+            return ResponseEntity.ok(teacher);
+
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     // Aktualisiert einen Lehrer
